refactor(client): extract the Edges tab into a self-loading OnPush component (A17.9b-6f)
Third record tab out of the shell, following the memories pattern. edges-tab.component
owns the edge create form, inline edit, delete, and its own text/semantic search (via
store.edgeSearch/edgeSearchMode, like memories) + type-tag filter + pagination + loader;
self-loads via a spaceId effect.

Edge-specific behaviour preserved (pinned by the relocated 6b cases): create AND
inline-edit strip empty optional properties via the edge schema; deleteEdge does NOT
refresh the space stats, so it emits no `mutated` (the asymmetry with memory/entity).
The edge from/to endpoint pickers (pickEdgeFrom/pickEdgeTo, on the shell since the
picker split) moved into the tab with edgeForm.

The 6b createEdge/deleteEdge cases + the two edge-endpoint tests relocated to
edges-tab.component.spec.ts. Removing the last of memories/entities/edges made
PropertiesViewComponent/PropertiesEditorComponent unused in the shell (chrono/filemeta
have no schema-property editor) — dropped from its imports.

brain.component.ts 1762 -> 1402. Client suite 197 -> 200. tsc --noUnusedLocals clean;
prod AOT build warning-free bar the pre-existing qrcode CJS bailout.
